chore: remove InMemoryState trait
